realign file structure
