+++ app/src/main/java/com/cid/bot/models.kt
@@ -1,5 +1,5 @@
 package com.cid.bot
 
-data class Muser(val id: Int?, val username: String, val gender: Int, val birthdate: String?)
 
-data class Message(val id: Int?, val sender: String?, val receiver: String?, val text: String, val created: String? = null)
+
+
