+++ app/src/main/res/layout/activity_profile.xml
@@ -81,7 +81,7 @@
                             android:hint="yyyy-MM-dd"
                             android:inputType="date"
                             android:singleLine="true"
-                            android:text="@{viewModel.text}"
+                            android:text="@{viewModel.muser.birthdate}"
                             app:layout_constraintBottom_toBottomOf="parent"
                             app:layout_constraintEnd_toEndOf="parent"
                             app:layout_constraintTop_toTopOf="parent" />
@@ -128,6 +128,7 @@
                             android:layout_marginEnd="8dp"
                             android:layout_marginBottom="8dp"
                             android:entries="@array/genderTypes"
+                            android:selectedItemPosition="@{viewModel.muser.gender}"
                             app:layout_constraintBottom_toBottomOf="parent"
                             app:layout_constraintEnd_toEndOf="parent"
                             app:layout_constraintTop_toTopOf="parent" />
